pyproject.toml
[project]
name = "sprawl-index"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4"
dependencies = [
    "geopandas (>=1.0.1,<2.0.0)",
    "rasterio (>=1.4.3,<2.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "openrouteservice (>=2.3.3,<3.0.0)",
    "fiona (>=1.10.1,<2.0.0)",
    "rasterstats (>=0.20.0,<0.21.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
